pyproject.toml
[project]
name = "base-fast-sqla"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "fastapi[all]>=0.115.12",
    "jose>=1.0.0",
    "pydantic>=2.11.4",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "sqlalchemy[asyncio]>=2.0.40",
]
